Renames the separated custom messages files to prevent conflict.
