+++ soh/soh/Enhancements/randomizer/randomizer_custom_messages.cpp
@@ -1,4 +1,4 @@
-#include "custom_messages.h"
+#include "randomizer_custom_messages.h"
 #include <variables.h>
 
 using namespace std::literals::string_literals;
